+++ lib/components/screens/victoryscreen.dart
@@ -26,10 +26,10 @@ class VictoryScreen extends Component
     add(victory);
 
     final style = TextStyle(
-        color: BasicPalette.black.color, fontSize: 100, fontFamily: 'Komika');
+        color: BasicPalette.black.color, fontSize: 100, fontFamily: 'Poland');
 
     final style_small = TextStyle(
-        color: BasicPalette.black.color, fontSize: 30, fontFamily: 'Komika');
+        color: BasicPalette.black.color, fontSize: 30, fontFamily: 'Poland');
 
     TextPaint textPaint = TextPaint(style: style);
     TextPaint textPaint_small = TextPaint(style: style_small);
@@ -49,7 +49,7 @@ class VictoryScreen extends Component
         textRenderer: textPaint_small,
         anchor: Anchor.center);
 
-    stats.position = Vector2(game.size.x / 2, scale.y * 700);
+    stats.position = Vector2(game.size.x / 2, scale.y * 690);
     stats.scale = scale;
 
     add(stats);